[Refactor] Refactor authentication code and rename 'User' entity
Identifiers and parameter type definitions in the authentication module were improved for better readability and robustness. Specifically, the 'User' class was renamed to 'UserEntity' to more accurately reflect its nature. This change is reflected in various controllers and services where the 'UserEntity' is used.

Additionally, User Id's type was updated from 'string' to more precise 'UUID' in certain contexts. Changing parameter types will clarify what types of parameters are expected and help prevent potential type errors.

The parsing of User Id from string to integer in the UserController's methods where User Id was used as a route parameter is also updated, for better validation of route parameters.

Overall, these refactorings and renaming are hoped to make the code more understandable, robust and easy to maintain.