[Feature] Add logout functionality in user service
Removed unused 'filter' import from user.service.ts as it was not being utilized anywhere in the file. In user.controller.ts, added a new endpoint for user logout that goes on to call the newly constructed user logout service with the passed user id. This functionality was necessary for handling user sessions.