[Refactor] Refactor user service and controller to improve data handling
In this commit, the methods of the user service and controller were updated to better handle user data. The return type of the findOne and updateUserData methods were changed to Promise<UserResponse>. A new method, findOneUser, was added to the user service to handle the conversion of user entity to user response internally, improving data encapsulation. Hashing of the user's password was added in updateUserDto to enhance security. The controller was updated to use these changes.