[Feature] Add CORS setup and configurations
CORS related setups and configurations have been added to the config and main application module files. This change was made to enhance the security by controlling the web sites that can interact with our server, and also to handle cross-origin issues. Further, the environmental variables for CORS configuration have been updated in the .env file.