[Feature] Add Helmet middleware for security enhancements
Helmet library has been added to the project for the purpose of securing the HTTP requests. It adds various HTTP headers that enhance the security of the application. The `helmet` is now included in package.json and package-lock.json as a dependency and is being utilized in main.ts as middleware.