[Feature] Add validation for email and password length in DTOs
The `register.dto.ts` and `create-user.dto.ts` files were updated to include length validation for the email and password fields. The 'MaxLength' and 'MinLength' decorators from 'class-validator' were utilised for this purpose. Now, the email can't be longer than 64 characters, and the password must be between 6 and 16 characters. This change was necessary to enforce consistent data format and ensure security.